+++ components/script/dom/bindings/codegen/Bindings.conf
@@ -140,7 +140,7 @@ DOMInterfaces = {
     'inRealms': ['WatchAdvertisements'],
 },
 
-'XR': {
+'XRSystem': {
     'inRealms': ['SupportsSessionMode', 'RequestSession'],
 },
