+++ components/script/dom/mod.rs
@@ -573,7 +573,6 @@ pub mod xmlhttprequest;
 pub mod xmlhttprequesteventtarget;
 pub mod xmlhttprequestupload;
 pub mod xmlserializer;
-pub mod xr;
 pub mod xrframe;
 pub mod xrinputsource;
 pub mod xrinputsourcearray;
@@ -586,6 +585,7 @@ pub mod xrrigidtransform;
 pub mod xrsession;
 pub mod xrsessionevent;
 pub mod xrspace;
+pub mod xrsystem;
 pub mod xrtest;
 pub mod xrview;
 pub mod xrviewerpose;
