+++ components/script/dom/webidls/XRSystem.webidl
@@ -2,9 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-// https://immersive-web.github.io/webxr/#xr-interface
+// https://immersive-web.github.io/webxr/#xrsystem-interface
 [SecureContext, Exposed=Window, Pref="dom.webxr.enabled"]
-interface XR: EventTarget {
+interface XRSystem: EventTarget {
   // Methods
   Promise<boolean> isSessionSupported(XRSessionMode mode);
   Promise<XRSession> requestSession(XRSessionMode mode, optional  XRSessionInit parameters = {});
@@ -15,7 +15,7 @@ interface XR: EventTarget {
 
 [SecureContext]
 partial interface Navigator {
-  [SameObject, Pref="dom.webxr.enabled"] readonly attribute XR xr;
+  [SameObject, Pref="dom.webxr.enabled"] readonly attribute XRSystem xr;
 };
 
 enum XRSessionMode {
@@ -29,7 +29,7 @@ dictionary XRSessionInit {
   sequence<any> optionalFeatures;
 };
 
-partial interface XR {
+partial interface XRSystem {
   // https://github.com/immersive-web/webxr-test-api/
   [SameObject, Pref="dom.webxr.test"] readonly attribute XRTest test;
 };
