+++ tests/wpt/metadata/webxr/idlharness.https.window.js.ini
@@ -131,25 +131,7 @@
   [XRSystem interface: operation isSessionSupported(XRSessionMode)]
     expected: FAIL
 
-  [XRSystem interface: existence and properties of interface prototype object]
-    expected: FAIL
-
-  [XRSystem interface: existence and properties of interface object]
-    expected: FAIL
-
-  [XRSystem interface object name]
-    expected: FAIL
-
   [XRSystem interface: attribute ondevicechange]
-    expected: FAIL
-
-  [XRSystem interface: existence and properties of interface prototype object's @@unscopables property]
-    expected: FAIL
-
-  [XRSystem interface: existence and properties of interface prototype object's "constructor" property]
-    expected: FAIL
-
-  [XRSystem interface object length]
     expected: FAIL
 
   [XRSystem interface: operation requestSession(XRSessionMode, XRSessionInit)]
