+++ tests/wpt/mozilla/meta/MANIFEST.json
@@ -13863,7 +13863,7 @@
      ]
     ],
     "interfaces.html": [
-     "bb6adbde0b3452e6ea8f1db44941a08dbe236774",
+     "163ab416b1627e07347b83aa210712364182343a",
      [
       null,
       {}
