+++ tests/wpt/mozilla/tests/mozilla/interfaces.html
@@ -263,7 +263,6 @@ test_interfaces([
   "XMLHttpRequestEventTarget",
   "XMLHttpRequestUpload",
   "XMLSerializer",
-  "XR",
   "XRFrame",
   "XRInputSource",
   "XRInputSourceArray",
@@ -275,6 +274,7 @@ test_interfaces([
   "XRSession",
   "XRSessionEvent",
   "XRSpace",
+  "XRSystem",
   "XRView",
   "XRViewerPose",
   "XRViewport",
